Reduce overhead of profiler operations when not needed.
